Ajout game

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -21,8 +21,8 @@ class CreateUsersTable extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->decimal('step_vid')->default(0);
-            $table->decimal('step_story')->default(0);
+            $table->bigInteger('step_vid')->default(0);
+            $table->bigInteger('step_story')->default(0);
 
             $table->rememberToken();
             $table->timestamps();
